Updated user, added authorization to every repos, added user service for token accesing

+++ Projekt_inz_backend/Controllers/DndClassController.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
@@ -6,6 +7,7 @@ using Projekt_inz_backend.Dto;
 using Projekt_inz_backend.Interfaces;
 using Projekt_inz_backend.Models;
 using Projekt_inz_backend.Repository;
+using Projekt_inz_backend.Services.UserServices;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -17,39 +19,41 @@ namespace Projekt_inz_backend.Controllers
     {
         private readonly IDndClassRepository _dndclassrepos;
         private readonly IMapper _mapper;
+        private readonly IUserService _userservice;
 
-        public DndClassController(IDndClassRepository classrepos, IMapper mapper)
+        public DndClassController(IDndClassRepository classrepos, IMapper mapper, IUserService userservice)
         {
             _dndclassrepos = classrepos;
             _mapper = mapper;
+            _userservice = userservice;
         }
         // GET: api/dndclass
-        [HttpGet]
+        [HttpGet, AllowAnonymous]
         public IActionResult Get()
         {
             return Ok(_mapper.Map<List<DndClassDto>>(_dndclassrepos.getDndClasses()));
         }
 
         // GET api/dndclass/id/1
-        [HttpGet("id/{id}")]
+        [HttpGet("id/{id}"), AllowAnonymous]
         public IActionResult Get(int id)
         {
             return Ok(_mapper.Map<DndClassDto>(_dndclassrepos.getDndClass(id)));
         }
-        [HttpGet("name/{name}")]
+        [HttpGet("name/{name}"), AllowAnonymous]
         public IActionResult GetByName(string name)
         {
             return Ok(_mapper.Map<List<DndClassDto>>(_dndclassrepos.GetDndClass(name)));
         }
         //GET api/dndclass/spells/1
-        [HttpGet("spells/{id}")]
+        [HttpGet("spells/{id}"), AllowAnonymous]
         public IActionResult GetSpells(int id)
         {
             return Ok(_mapper.Map<List<SpellDto>>(_dndclassrepos.GetClassSpells(id)));
         }
         // POST api/database
-        [HttpPost]
-        public IActionResult CreateDndClass(int ownerid, [FromBody] DndClassDto dndClass)
+        [HttpPost, Authorize(Roles = "user,admin")]
+        public IActionResult CreateDndClass([FromBody] DndClassDto dndClass)
         {
             if (dndClass == null)
             {
@@ -57,7 +61,7 @@ namespace Projekt_inz_backend.Controllers
             }
             var dndClassMap = _mapper.Map<DndClass>(dndClass);
 
-            if (!_dndclassrepos.CreateDndClass(ownerid, dndClassMap))
+            if (!_dndclassrepos.CreateDndClass(_dndclassrepos.GetUserIdByName(_userservice.GetName()), dndClassMap))
             {
                 ModelState.AddModelError("", "Cos poszlo nie tak z zapisem");
                 return StatusCode(500, ModelState);
@@ -66,29 +70,42 @@ namespace Projekt_inz_backend.Controllers
         }
 
         // PUT api/database
-        [HttpPut]
+        [HttpPut, Authorize(Roles = "user,admin")]
         public IActionResult UpdateDndClass([FromBody] DndClassDto dndClass)
         {
-            var dndClassMap = _mapper.Map<DndClass>(dndClass);
-            if (!_dndclassrepos.UpdateDndClass(dndClassMap))
+            if (_dndclassrepos.GetOwnerId(dndClass.classID) == _dndclassrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
-                return StatusCode(500, ModelState);
+                var dndClassMap = _mapper.Map<DndClass>(dndClass);
+                if (!_dndclassrepos.UpdateDndClass(dndClassMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do zmiany tego obiektu");
+            return StatusCode(403, ModelState);
         }
 
         // DELETE api/database
-        [HttpDelete]
+        [HttpDelete, Authorize(Roles = "user,admin")]
         public IActionResult DeleteDndClass(DndClassDto dndClass)
         {
-            var dndClassMap = _mapper.Map<DndClass>(dndClass);
-            if (!_dndclassrepos.DeleteDndClass(dndClassMap))
+            if (_dndclassrepos.GetOwnerId(dndClass.classID) == _dndclassrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
-                return StatusCode(500, ModelState);
+                var dndClassMap = _mapper.Map<DndClass>(dndClass);
+                if (!_dndclassrepos.DeleteDndClass(dndClassMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do usuniecia tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
     }
 }
