+++ Projekt_inz_backend/Controllers/RaceController.cs
@@ -1,10 +1,12 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using Projekt_inz_backend.Data;
 using Projekt_inz_backend.Dto;
 using Projekt_inz_backend.Interfaces;
 using Projekt_inz_backend.Models;
+using Projekt_inz_backend.Services.UserServices;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -16,34 +18,36 @@ namespace Projekt_inz_backend.Controllers
     {
         private readonly IRaceRepository _racerepos;
         private readonly IMapper _mapper;
+        private readonly IUserService _userservice;
 
         // GET: api/<RaceController>
 
-        public RaceController(IRaceRepository racerepos, IMapper mapper)
+        public RaceController(IRaceRepository racerepos, IMapper mapper, IUserService userservice)
         {
             _racerepos = racerepos;
             _mapper = mapper;
+            _userservice = userservice;
         }
-        [HttpGet]
+        [HttpGet, AllowAnonymous]
         public IActionResult Get()
         {
             return Ok(_mapper.Map<List<RaceDto>>(_racerepos.GetRaces()));
         }
 
         // GET api/<RaceController>/5
-        [HttpGet("id/{id}")]
+        [HttpGet("id/{id}"), AllowAnonymous]
         public IActionResult Get(int id)
         {
             return Ok(_mapper.Map<RaceDto>(_racerepos.GetRace(id)));
         }
-        [HttpGet("name/{name}")]
+        [HttpGet("name/{name}"), AllowAnonymous]
         public IActionResult GetByName(string name)
         {
             return Ok(_mapper.Map<List<RaceDto>>(_racerepos.GetRace(name)));
         }
         // POST api/<RaceController>
-        [HttpPost]
-        public IActionResult CreateRace(int ownerId, [FromBody] RaceDto race)
+        [HttpPost, Authorize(Roles = "user,admin")]
+        public IActionResult CreateRace([FromBody] RaceDto race)
         {
             if (race == null)
             {
@@ -51,7 +55,7 @@ namespace Projekt_inz_backend.Controllers
             }
             var raceMap = _mapper.Map<Race>(race);
 
-            if (!_racerepos.CreateRace(ownerId, raceMap))
+            if (!_racerepos.CreateRace(_racerepos.GetUserIdByName(_userservice.GetName()), raceMap))
             {
                 ModelState.AddModelError("", "Cos poszlo nie tak z zapisem");
                 return StatusCode(500, ModelState);
@@ -60,29 +64,42 @@ namespace Projekt_inz_backend.Controllers
         }
 
         // PUT api/<RaceController>/5
-        [HttpPut]
+        [HttpPut, Authorize(Roles = "user,admin")]
         public IActionResult UpdateRace([FromBody] RaceDto updatedRace)
         {
-            var raceMap = _mapper.Map<Race>(updatedRace);
-            if (!_racerepos.UpdateRace(raceMap))
+            if (_racerepos.GetOwnerId(updatedRace.raceID) == _racerepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
-                return StatusCode(500, ModelState);
+                var raceMap = _mapper.Map<Race>(updatedRace);
+                if (!_racerepos.UpdateRace(raceMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do zmiany tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
 
         // DELETE api/<RaceController>/5
-        [HttpDelete]
+        [HttpDelete, Authorize(Roles = "user,admin")]
         public IActionResult DeleteRace(RaceDto race)
         {
-            var raceMap = _mapper.Map<Race>(race);
-            if (!_racerepos.DeleteRace(raceMap))
+            if (_racerepos.GetOwnerId(race.raceID) == _racerepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
-                return StatusCode(500, ModelState);
+                var raceMap = _mapper.Map<Race>(race);
+                if (!_racerepos.DeleteRace(raceMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do usuniecia tego obiektu");
+            return StatusCode(403, ModelState);
         }
     }
 }
